+++ src/main/java/com/library/libraryapi/DAO/FileRepository.java
@@ -1,0 +1,7 @@
+package com.library.libraryapi.DAO;
+
+import com.library.libraryapi.Model.File;
+import org.springframework.data.repository.CrudRepository;
+
+public interface FileRepository extends CrudRepository<File, Long> {
+}