+++ src/main/java/com/library/libraryapi/Model/File.java
@@ -1,0 +1,34 @@
+package com.library.libraryapi.Model;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.validation.constraints.NotNull;
+
+@Entity
+public class File {
+
+    @Id
+    @GeneratedValue
+    @NotNull
+    private long id;
+
+    @NotNull
+    private byte[] file;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public byte[] getFile() {
+        return file;
+    }
+
+    public void setFile(byte[] file) {
+        this.file = file;
+    }
+}